A bunch of parts working now
The add/delete button groups and the add buttons are now working. Still need to get the button details and the delete button to work. The delete button group will leave orphan buttons behind. I'm wondering if there is an advantage to leaving it that way so that if you delete the button group you could create another one and reattach the buttons to it. Would need to have an "Orphaned Buttons" group that displays on the settings page to deal with that.

+++ settings.php
@@ -44,7 +44,7 @@
               <td>' . $buttons . '</td>
               <td>
                 <a href="add_button.php?button_group_id=' . $row['button_group_id'] . '">Add Button</a><br>
-                <a href="delete_button_group.php?button_group_id=' . $row['button_group_id'] . '">Delete Group</a>
+                <a href="delete_button_group.php?button_group_id=' . $row['button_group_id'] . '" onclick="return confirm(\'Are you sure you want to delete this button group?\');">Delete Group</a>
               </td>
             </tr>
           ';
